% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Add transparency to colors}
\usage{
add.alpha(col, alpha = 1)
}
\description{
This function allows you to add transparancy to colors.
}
\examples{
add.alpha("black", 0.5)
}
\keyword{alpha}
\keyword{colors,}
\keyword{transparency,}
