% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice.lm.anova.plot.R
\name{nice.lm.anova.plot}
\alias{nice.lm.anova.plot}
\title{Plot nice LM plots}
\usage{
df <- data.frame
y <- string with the name of the y element in the formula. Column of data.
x <- string with the name of the x element in the formula. Column of data or string with the formula of the linear model.
}
\description{
This function plots descriptive, diagnostic and the results of anova on linear model. If a string of formula is provided formula instead of a single variable (i.e., "rain + heat" instead of  "heat"), the last element is plotted, although lm is done with the string formula. This function also return the fitted model and the anova resutls. Elements are in a list.
}
\examples{
nice.lm.anova.out(alpha, "S", "Intervention.timepoint")
}
\keyword{anova,}
\keyword{ggplot,}
\keyword{linear}
\keyword{lm}
\keyword{model,}
\keyword{nice}
\keyword{plot,}
