% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice.lm.anova.res.R
\name{nice.lm.anova.res}
\alias{nice.lm.anova.res}
\title{Plot nice LM plots}
\usage{
It returns a list, where the first element is a ggplot/ggarrange.
df <- data.frame
formula <- formula of the model. if string, use as.formula(string)
}
\description{
This function plots descriptive, diagnostic and the results of anova on linear model. If a string of formula is provided formula instead of a single variable (i.e., "rain + heat" instead of  "heat"), the last element is plotted, although lm is done with the string formula. This function also return the fitted model and the anova resutls. Elements are in a list.
}
\details{
At the moment it does not support functions with logs or transformations in the formula. transform the column in the dataframe, instead.
}
\examples{
#get results
data(mtcars)
f <- as.formula("mpg ~ cyl")
res <- nice.lm.anova.res(formula = f, data = mtcars)
res$plot
}
\keyword{anova,}
\keyword{ggplot,}
\keyword{linear}
\keyword{lm}
\keyword{model,}
\keyword{nice}
\keyword{plot,}
