% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.filler.R
\name{tax.filler}
\alias{tax.filler}
\title{Fill taxonomy based on previous hierarchy}
\usage{
tax.filler(df, start, end)
}
\description{
This function gets taxonomy df, each tax level per column, and fill the empty taxonomy levels based on the higher taxonomy hierarchy. Empty taxonomy are either NA or empty strings. Data frame has to be of character type and not factors.
}
\examples{
tax.filler(tax, 3,6)
}
\keyword{NA,}
\keyword{fill}
\keyword{taxonomy,}
