% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.filler.R
\name{tax.filler}
\alias{tax.filler}
\title{Fill taxonomy based on previous hierarchy}
\usage{
tax.filler(df, start, end)
}
\description{
This function gets taxonomy df, each tax level per column, and fill the empty taxonomy levels based on the higher taxonomy hierarchy. Empty taxonomy are either NA or empty strings.
}
\examples{
tax.filler(tax, 3,6)
}
\keyword{NA,}
\keyword{fill}
\keyword{taxonomy,}
